ai-model/requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.0
matplotlib==3.7.2
pandas==2.0.3